refactor: replace Prisma Employee model
