+++ src/types/employee.ts
@@ -1,0 +1,7 @@
+type Employee = {
+  email: string
+  firstName: string
+  lastName: string
+}
+
+export type { Employee }
